Refactor: Move SDL3 classes to "Imports" namespace
Updated SDL3 classes to the "Imports" namespace and changed their
accessibility from public to internal.

This improves code organization and encapsulation by clearly separating
imported functionality.
