+++ Open.CommandAndConquer.Sdl3/src/Imports/SDL_clipboard.cs
@@ -22,9 +22,9 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Runtime.InteropServices.Marshalling;
 
-namespace Open.CommandAndConquer.Sdl3;
+namespace Open.CommandAndConquer.Sdl3.Imports;
 
-public static partial class SDL3
+internal static partial class SDL3
 {
     [NativeMarshalling(typeof(SafeHandleMarshaller<SDL_ClipboardData>))]
     public sealed class SDL_ClipboardData : SafeHandle
