+++ Open.CommandAndConquer.Sdl3/src/Imports/SDL_iostream.cs
@@ -22,9 +22,9 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Runtime.InteropServices.Marshalling;
 
-namespace Open.CommandAndConquer.Sdl3;
+namespace Open.CommandAndConquer.Sdl3.Imports;
 
-public static partial class SDL3
+internal static partial class SDL3
 {
     public enum SDL_IOStatus
     {
