+++ Open.CommandAndConquer.Sdl3/src/Imports/SDL_rect.cs
@@ -21,9 +21,9 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Runtime.InteropServices.Marshalling;
 
-namespace Open.CommandAndConquer.Sdl3;
+namespace Open.CommandAndConquer.Sdl3.Imports;
 
-public static partial class SDL3
+internal static partial class SDL3
 {
     [StructLayout(LayoutKind.Sequential)]
     public struct SDL_Point
